--- src/main/java/david/buzas/whattoeat/MealCategory.java
@@ -1,9 +1,0 @@
-package david.buzas.whattoeat;
-
-import lombok.Data;
-
-@Data
-public class MealCategory {
-    private String key;
-    private String displayName;
-}
