--- src/main/java/david/buzas/whattoeat/MealRepository.java
@@ -1,8 +1,0 @@
-package david.buzas.whattoeat;
-
-public class MealRepository extends ItemRepository<Meal> {
-
-    public MealRepository() {
-        super(Meal.class, new JsonFileItemManager<>(Meal.class, "user-meals.json"));
-    }
-}
