--- src/main/java/david/buzas/whattoeat/MealType.java
@@ -1,9 +1,0 @@
-package david.buzas.whattoeat;
-
-import lombok.Data;
-
-@Data
-public class MealType {
-    private String key;
-    private String displayName;
-}
