+++ src/main/java/david/buzas/whattoeat/entities/Entity.java
@@ -1,4 +1,4 @@
-package david.buzas.whattoeat;
+package david.buzas.whattoeat.entities;
 
 import java.util.UUID;
